flood-guardian-server/requirements.txt
numpy==1.18.1
flask==1.1.2
scikit-learn==0.22.2.post1
pandas==1.0.1
requests==2.25.1
ibm-watson==5.1.0
python-dotenv==0.15.0
requests-cache==0.5.2
shapely==1.7.1
pytest==6.2.4